wip: #60 moved next-auth options to lib file
