+++ apps/issues/lib/auth.ts
@@ -1,0 +1,25 @@
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import type { NextAuthOptions } from 'next-auth';
+import GithubProvider from 'next-auth/providers/github';
+
+import { prisma } from './db';
+import { env } from '../server/env';
+
+export const authOptions: NextAuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  callbacks: {
+    async session({ session, user }) {
+      if (session.user) {
+        session.user.id = user.id;
+        session.user.settings = user.settings;
+      }
+      return session;
+    },
+  },
+  providers: [
+    GithubProvider({
+      clientId: env.GITHUB_CLIENT_ID,
+      clientSecret: env.GITHUB_CLIENT_SECRET,
+    }),
+  ],
+};
