+++ apps/issues/pages/api/auth/[...nextauth].ts
@@ -1,27 +1,5 @@
-import { PrismaAdapter } from '@next-auth/prisma-adapter';
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GithubProvider from 'next-auth/providers/github';
+import NextAuth from 'next-auth';
 
-import { prisma } from '../../../lib/db';
-import { env } from '../../../server/env';
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  callbacks: {
-    async session({ session, user }) {
-      if (session.user) {
-        session.user.id = user.id;
-        session.user.settings = user.settings;
-      }
-      return session;
-    },
-  },
-  providers: [
-    GithubProvider({
-      clientId: env.GITHUB_CLIENT_ID,
-      clientSecret: env.GITHUB_CLIENT_SECRET,
-    }),
-  ],
-};
+import { authOptions } from '../../../lib/auth';
 
 export default NextAuth(authOptions);
