+++ apps/issues/server/router/context.ts
@@ -5,8 +5,8 @@ import {
 import * as trpc from '@trpc/server';
 import * as trpcNext from '@trpc/server/adapters/next';
 
+import { authOptions as nextAuthOptions } from '../../lib/auth';
 import { prisma } from '../../lib/db';
-import { authOptions as nextAuthOptions } from '../../pages/api/auth/[...nextauth]';
 
 type CreateContextOptions = {
   session: Session | null;
